Grosse amélioration de l'affichage des triangles en fonction de la normale pour la décroissance de la lumière.

+++ roam.h
@@ -6,9 +6,9 @@ typedef struct Vertex {
        int x;
        int y;
        int z;
-       int xNormal;
-       int yNormal;
-       int zNormal;
+       float xNormal;
+       float yNormal;
+       float zNormal;
        /* Ajouter des champs ici. */
 } Vertex;
